Fase 2/Evidencias Proyecto/Evidencias de sistema de Aplicación/requirements.txt
asgiref==3.8.1
Brotli==1.1.0
cffi==1.16.0
chardet==5.2.0
cssselect2==0.7.0
Django==5.0.6
djangorestframework==3.15.2
fonttools==4.53.0
html5lib==1.1
Jinja2==3.1.4
MarkupSafe==2.1.5
pdfkit==1.0.0
pillow==10.4.0
psycopg2==2.9.10
pycparser==2.22
pydyf==0.10.0
pyphen==0.15.0
reportlab==4.2.2
six==1.16.0
sqlparse==0.5.0
tinycss2==1.3.0
tzdata==2024.1
weasyprint==62.3
webencodings==0.5.1
wkhtmltopdf==0.2
zopfli==0.2.3
